Add lint check for PPD purpose string content.
When a purpose string is required to be declared for a permission, the
string resource must reference strings that are not blank and no more
than 300 characters long as these are shown to users.

Added a separate issue for this check in case there are performance
issues and this issue needs to be suppressed without sacrificing other
purpose declaration lint checks.

Bug: 442571488
Test: PurposeDeclarationDetectorTest
Change-Id: If94fbefd448c6912718626cd4fd3314a5f075659