forcefully clean output folders when running test engines.
Because test engines are running in a separate process, Gradle does not
have the ability to monitor what is written in the folder annotated with
@OutputDirectory so it leaves them untouched at the next execution.
This is a bit surprising but I will clear with Gradle folks that it is
their intented policy.
In the meantime, forcefully clean the results and coverage dirs since
the task is not incremental anyway.

Test: Added a regression test
Bug: N/A
Change-Id: I7eeabc9f01627c60892c9837b8800d751c4af1bc