Intern common values with frequent calls to `equals` and `hashCode`
Within an analysis, class names, method names, and symbolic invocations
are finite value sets (assuming no bug that fails to summarize the
symbols), and they are frequently compared for equality and queried for
hash codes.

This change interns these values, and replaces logical equality and
`hashCode` with referential equality and `System.identityHashCode` where
safe. The change is measured to result in a 4-7x speed-up in large
projects.

Bug: n/a
Test: existing + new
Change-Id: I8ba248c9a09a512720b45a88401af437e8ee07cb