Add utils for assuming types and effects in base environment
The analysis gets types and effects of methods from other modules
either from having analyzed them, or admitting them as axioms in a
base environment.

This change adds a DSL for assuming those (`ResultTableBuilder`), with
`given` for building a "mono" signature, and `forAll` for introducing
a type parameter. Further, because both the meta program and the
object program have access to the same Java and Kotlin utils, we use
`kotlin.reflect.*` to reduce boilerplates and mistakes compared to
manually crafting those names and signatures.

Bug: n/a
Test: future CL
Change-Id: I5b7c61610630e7893074624b2113cfa3934925ae