Add dependencies and tweak some versions for arch sample template
Add many dependencies that the agent was adding for tier-2 and tier-4
apps.

Tweak some version numbers.

Add the internet permission to the manifest.

Note: the versions used here are not always the latest versions because
the version of Kotlin being used is 2.0.21, and some of the latest
versions are not compatible with Kotlin 2.0.21.

Bug: 449546776
Test: manual
Change-Id: I2d7261e4689bc9e21fa43f724104e1c67813cdd5