[fused lib] Fix minSdk when reciever used in kts
The GradleRule test framework only writes build files in Groovy. This
means when minSdk block was added in
I0dcc78ea1eb71029d36ccf0f7b741f833e960519, it was not tested for kts
build files. This changes will correctly configure the action for the
minSdk reciever in kts files.

Bug: 383124910
Test: manual
Change-Id: I587d90652c062aeb958b9128825b233ef4477800