Remove AppPreBuild Task
Reason for removal now is that in AGP 9.0 android.useAndroidX is enabled
and android.dependency.useConstraints is disabled - we see failed test:
b/440264272 due to versions of dependencies not being aligned between
runtime and compile classpath.

Instead of making the check more lenient (compare group and name,
without version) we opt for removal of the check.

Bug: 452612252
Test: N/A
Change-Id: Iacbe8559bdfd538b77f3fa3495b72261ffaf1099