Add an Option to Not Hook Framework APIs
With the ability of hooking custom AndroidX Sqlite Drivers, we open the possibility
for app to hook a driver that delegates to the Framework API either directly or indirectly.

For example, an app may implement a driver that encrypts data but uses the Framework API to read/write the encrypted data.
To allow this type of driver to be useful, we have to avoid hooking the framework API directly.

Bug: 449501590
Test: Added
Change-Id: I9f19d0ca4397cf5bcf06ee702d65400093ae9a40