Pass serial number to createRunnerConfigProtoForLocalDevice
This simplifies the interface of `createRunnerConfigProtoForLocalDevice`
by removing its dependency on the `DeviceConnector` object and requiring
only the device serial number string.

As part of this change, the logic to determine the on-device additional
test output directory is moved from the factory to the caller
(`UtpTestRunner`). This path is now passed as a new
'additionalTestOutputOnDeviceDir' argument.

This refactoring is a preparatory step for moving `UtpConfigFactory`
inside `RunUtpWorkAction`.

Bug: 450407553
Test: exists
Change-Id: Icaec1b9cc58e27176c02a3ee069be38f539ee151