Remove retry logic for EmulatorTimeoutException
The `EmulatorTimeoutException` was previously thrown by the
`android-device-provider-gradle` UTP plugin. This plugin is no longer in
use, as it was replaced by a default device provider (b/431284686).

Since the exception is no longer thrown, the corresponding retry logic
in `ManagedDeviceTestRunner` is obsolete and has been removed, along
with its helper functions (`hasEmulatorTimeoutException`) and related
unit tests.

Bug: 450407553
Test: exists
Change-Id: Iefdfd40d3fa9f9427588d7051e07f4d7073931c1