Fix issue in GradleDetector related to prefix ranges
If the separator character was not a '.' (as in e.g. `21.+`) but
instead some other acceptable separator character, such as '-', '_',
the absence of a character or even (hilariously) '+', the resulting
RichVersion associated with a dependency declaration would not have
the same string form as the original, which would cause the ReplaceFix
for various issues (such as the fix for newer versions being
available) not matching the text in the editor, leading to an IDE
Error.

The fix in this CL is not complete, because there are multiple
different ways of spelling various other similar constructs: the
universal range can be spelt "+" but also "[,]" or "(,)"; maven ranges
can be spelt with ']' in place of '(' and/or '[' in place of ')'.

We will also need to decide how to handle equality: should
VersionRanges with identical semantics but different printed
representation be considered `equals`?  (The existing precedent for
Version suggests "yes", but it's not clear that this was a conscious
decision)

Bug: 453765900 (in the logs)
Test: added, adjusted
Change-Id: Ib6ef5460c79b0afc744199cd3c47b1936becd4eb