Preserve the input identifier in all uses of the #parse() factory
There are a lot of different ways of spelling
otherwise-semantically-identical Gradle Version Ranges.  Since those
spellings might come from end-user input, it's important that we can
reproduce what we were given, so that text substitution (for example)
can happen without error.

If we parse a VersionRange, we can simply keep what we parsed.  If we
construct a VersionRange programmatically, compute the most plausible
identifier, if any, and store that.

Bug: 453765900 (in the logs)
Test: added, adjusted
Change-Id: Ia94b00a7152d1f1da3e9903e41bc4b6ec3f8c901