initial work to support test suite sources processing.
Test Suites can have multiple sources, each with a single type. So
far, we only supported ASSETS based sources. This is the pre-work
to start supporting HOST_JAR (and eventually TEST_APK) source types.
Enhanced the source container for each source defined in a test suite.
The source container has it own set of dependencies (which can be declared
in the DSL) and Artifacts instance that will store the built
artifacts once compilation and other processing are added.

Relnote: N/A
Test: Some unit test for new interfaces
Bug: N/A
Change-Id: I556abc67a1645cb9a0fb5f04b0bc7b94cdc6a172