reworked how source files are expressed in model
instead of having fields that can be nullable
depending on the source type like assets and host jar,
use dedicated interface for each source type so can
precisely control which directories are available.
Because of serialization of the model, we cannot use a
single collection to pass these sources, so instead
use a dedicated collection for each type.

Test: Existing
Bug: N/A
Change-Id: Ib3c1c65aafd749292a4fa7a5a8675d8e5e548a51