Introduce hybrid test project fixtures
and convert LintDesugaringTest to new fixtures.

This test uses an on-disk test-project and therefore
could not be directly converted to the new fixture.

This change introduces a way for the GradleRule
test fixture to load project source files from the
disk, while describing the build files via the
GradleRule DSL.

This will allow us to have more complex projects
while still allowing us to maintain their build
files via a compiled DSL.

Bug: N/A
Test: existing tests
Change-Id: I962d897910a34053d0ddec456ab71087ada27008