Fix GradleBuild.withReversibleModifications
This did not work at all :( This was discovered
when converting ResPackagingTest.

This was implemented by reusing TemporaryProjectModification
but this was not really adapted. This also made things much
more complicated by creating a Reversible*ProjectFiles
that connected to the original API for TPM.

The new implementation is much simpler. The implementation
of Reversible*ProjectFiles is the one doing the recording
of the state in a way that is just a small layer on top
of the Direct*ProjectFiles implementation making things
much easier to maintain moving forward.

This also cleans up the event recording. While
the events described actuall changes (edit/remove/...)
only the first "change" was recorded in order to restore
the original state. The new mechanism is now recording
actual "original state" rather than "events" to make
it clearer what is going on.

Bug: N/A
Test: existing tests + new unit tests
Change-Id: If74dd2f0e7e02e887045b5befcdc89b13177b223