Handle gRPC CANCELLED status in UTP result listener
When the `onError` handler in `UtpTestResultListenerServer` receives a
`StatusRuntimeException` with a `CANCELLED` status, this change now
reports the entire test suite as `CANCELLED`.

This avoids logging a confusing generic error and provides a more
accurate test status. This situation typically occurs when a user aborts
the Gradle build, which cancels the gRPC stream.

Fixes: 456992735
Test: UtpTestResultListenerServerTest.recordTestResultEvent_CancelledError
Change-Id: I8221822fed89a8c7767de982d251e01493c66d4b