Fix some tests to not opt-out of newDsl/enableAppCompileTimeRClass
This commits adds more opt-out detection issues in
tests both for missing built-in kotlin opt-out and
newDsl opt-out based on used plugins and callbacks.
This is part of the GradleRule fixture only.

This also tests for extraneous opt-outs, and fails.
This is to guarantee that tests that can run with
the new default behaviors actually do so.

There are corner cases where this latter detection
can be wrong. In that case, there's a way to disable
it as an option on the GradleRule.

Based on these checks, the commit includes
fixes for several tests using the GradleRule, and
disables blanket opt-outs for newDsl, enableAppCompileTimeRClass
and enableLegacyVariantApi (which shouldn't be used
anyway). This also removed usage of disableBuiltInKotlin
in all GradleRule-based tests.
(several obsolete imports of this methods also had
to be fixed even if the tests were untouched)

Some tests do manually opt out of some behaviors
where needed, but it means that new tests will
use the new behaviors by default, and we can
now search of usages of the opt-out to fix
tests one by one.

This also fixes deprecation on the kotlin-android plugin
(instead of kotlin-jvm), and suppress deprecation warning
on some expected usages.

Bug: N/A
Test: N/A
Change-Id: I018a51e77d00171d01ba56f67089b5174f024147