Track more precise type for immutable local variables
The analysis was using the host type system's result for declared
variables, discarding the RHS's inferred type and effect.

For the case where there's no user declaration, it's clearly reasonable
that we should use the analysis's more precise result. But for the case
where the user explicitly declares a super-type, the analysis will also
break that user-established abstraction if it's a local immutable
variable, for practical reasons (e.g. most code is written before this
analysis exists). If the user wants a mutable variable that tracks
different function types with compatible effects, they will have to
add it to the explicit declaration.

(This change isn't addressing the known issue where we're not reporting
errors when a symbolic effect in an "exist" position fails a constraint)

Bug: 458451757
Test: new
Change-Id: I3de2dfec7f7420419e7b37d00494ec56e8f17b99