Split opt-out flags in old and new fixtures
Previous versions was done in the executor (for building
and syncing) and applied to both fixtures. With this
commit, the opt-out are now done differently for each
fixture.

We want to be able to fix tests using the new fixture
faster as we can implement various checks to help us
migrate.

A later commit is putting checks in places in GradleRule to
allow us to detect which tests are already compatible with the
new behavior, allowing us to convert them much faster.

The previous implementation in the executor injected the
opt-out as properties in the gradle command. While the
GradleTestProject fixture still does that, it's
configured in GradleTestProject instead of the generic
executor.

On the GradleRule side, the configuration is done
at the project level, in gradle.properties. This allows
for debugging the actual behavior via the project
created on disk. We still inject the suppress warning
property in the model builder as warning suppression
in gradle.properties does not impact the content of
SyncIssues.

Bug: N/A
Test: existing tests.
Change-Id: I94817f5d6aeabb3de7ab12e9568e17b0a436e87d