Added APIs to add dynamic directories to test suites.
Since test suites are not specific to a single variant, it does
not make sense to allow adding variant specific directories.
Therefore new APIs were added to the source declaration of the
test suites that will allow registering dynamically
generated directories.

The API is private for now and we will make them public
when we are more comfortable with the API surface.

The model is not quite correct at this point as
the static and generated folders are added to the same
BasicAndroidProject hierarchy. This will be reworked in the
next CL when static and generated folders will be respectively
added to the BasicAndroidProject and AndroidProject models.

Test: Added unit and integration tests
Bug: N/A
Relnote: N/A

Change-Id: I0164a94ac5a2c4b72e83664ce974b99c8dd7f8c7