Proper support for DSL extension in GradleRule
This fixes a test that relied on it for KTS
testing. An interesting aspect here is that
the plugin relied on doing println from
inside onVariants` and validating the content
of stdout and this is apparently not reliable
at all. While running tests locally, a lot
of flakiness could be observered. The fix
is to run actual tasks that output files
that can be validated.

This also fixes how we apply custom plugins
by creating the matching properties file
and doing a normal plugin block.

Fixes: 455844860
Test: fixed integration test and added some unit tests
Change-Id: Iae3eb049dab7bd48a27abc76da1524d4888bdb30