Remove `tailrec` in loop instantiating inductive summary
The loop for taking the fixpoint of an inductive summary isn't supposed
to take more than a few iterations, let alone go near stack overflow. If
there's a bug that prevents convergence, it's better to let the stack
overflow then catch and fix in unit tests or move on in production, as
opposed to letting it run for very long.

Bug: n/a
Test: n/a
Change-Id: Iae552513d11d7788f4782f5189fee178919fb36c