Clean up `syncAgpAndKgpSources` method for KMP modules
When tested with AGP 8.13.1, Gradle 8.13, and KGP 2.0.0 to 2.20.21,
the `syncAgpAndKgpSources` method in `kgpUtils.kt` does not actually do
anything in a `com.android.library` module that applies the
`org.jetbrains.kotlin.multiplatform` plugin.

The following line always returns null

  val kotlinConvention = plugins["kotlin"] ?: return null

This suggests that that code is probably obsolete, so this commit
removes that code to simplify the method.

Bug: 246910305
Test: Existing tests (refactoring only)
Change-Id: I5a46553d74eff9a031ac85961430d023698a5a1c