Add Maven modules for Lucene 9.12.
This will be bundled in aiplugin, rather than depending on the version
of Lucene bundled with IntelliJ. This will allow aiplugin to update its
version of Lucene independently, once we no longer rely on versions of
IntelliJ that bundle Lucene in the core classpath. This will also
prevent version-skew issues with Lucene libraries (e.g. queryparser)
that aiplugin depends on, but that are _not_ bundled with IJ.

Bug: 460307944
Test: N/A
Change-Id: Ib5cef430a9de7c639f2d5c1057fb45c7dd56a06a