WrongGradleMethodDetector: unwrap flexible types
If a type comes from Java and has unknown nullability
(e.g. ApplicationProductFlavor!) then the Kotlin compiler wraps it in
a KaFlexibleType which cannot be cast to KaClassType.

In WrongGradleMethodDetector we only care about the underlying class
name, so we can use the lower bound of the flexible type to get it.

This change fixes AndroidLintGradleTest.testWrongGradleMethod with
IntelliJ 2025.3 (which bundles a newer Kotlin compiler).

Bug: not filed
Test: AndroidLintGradleTest.testWrongGradleMethod with IJ 2025.3
Change-Id: I5befc7a66e7b6b5b1c9a95b03aa5570be195e55f