Adapt CmakeBasicProjectTest and NdkBuildBuildSettingsTest
Explicitly specify targetSdk (as the same as minSdk) in order not to
have to modify test files.  This trips a lint error, so suppress the
expired Sdk inspection that would otherwise fail.

Both these tests also use the old Variant API to assert some
not-very-interesting invariants (pun not intended).  Disable the
newDsl flag, but a relatively small piece of cleanup would be to find
some other way of testing what is currently tested in the
`applicationVariants.all` blocks, and un-disable the flag.

Nevertheless this partially unblocks running tests in a mode
reflecting AGP 9.x defaults.

Bug: none filed
Test: existing
Change-Id: Ia7d426204691f5a514b5b164c916c101511c4b89