Adapt some native/AGP integration tests
Explicitly specify targetSdk alongside minSdk.  Suppress the resulting
lint inspection.

It's possible that fixing the reported problem in the project (adding
an explicit "export" might have been simpler); on the other hand, that
can be done in the fullness of time, independently of the need to run
the tests with default AGP 9.x behaviours.

Bug: none filed
Test: existing
Change-Id: I2283ac62f6dd78bbac7323fa988dda5ecc4b77ae