Fix a wrong comparison in GradleVisitor
We need to check included file is within the project or not.
From the given kts file, analysis scope is bound to the kts module with
that specific kts file. Anything else, even within the project, will be
determined as if it's outside the project. For now, we stick to the
old-fashioned file path comparison.

Bug: 463283604
Test: presubmit
Change-Id: I00dfd030e6a2d30a65cc5b8dd589fc813c7a4b57