Make BootClasspathTest explicitly set newDsl to false
It's not clear to me why this test is testing what it is testing, but
it certainly appears to be the case that the new way of getting the
boot classpath does not work this early by design.  Add
cross-references to increase the chances that cleanup will happen once
USE_NEW_DSL is no longer an option.

Bug: none filed
Test: this is one
Change-Id: I7f6213b419ff4e65d56c151f2f88cc3c72c87c1d