Make Navigation AGP integration tests 9.x compatible
The new world defaults the targetSdk from the compileSdk, which is
implicit in these tests, so adjust the expected merged manifest
content appropriately.  Also, the manifest must have an explicit value
for the `android:exported` activity attribute, and consequently so
must the expected merged manifest.

In order to prepare for the new world while AGP_9_OPT_OUTS are being
applied, though, we have to override the override of
DEFAULT_TARGET_SDK_TO_COMPILE_SDK_IF_UNSET back to its 9.x value,
otherwise we fail the tests in the current setup.  Define and use a
custom executor in these tests.  (Remove the previous iteration of
this for `android.useAndroidX` in the test class @Before methods.)

Bug: none filed
Test: this is it
Change-Id: Ic6dfcfed58e7681c748e77b8afffde5ea072ddfa