Allow an test method to run with AGP-9.x defaults
Most of the test is applicable to both the old and the new world, but
the drive-by check that BaseExtension#bootClasspath doesn't break Sdk
downloading is only relevant for USE_NEW_DSL set to false.  Make it
so, and leave breadcrumbs for future maintainers on what to do when
USE_NEW_DSL no longer exists.

Bug: none filed
Test: test-only change
Change-Id: Ie611bc04d6dad02f921ce98168a5ffb91b531133