Address a couple of connected test AGP 9.x incompatibilities
Run ApiConnectedTest with USE_NEW_DSL set to false, for now.

Update targetProjectPath to assignment syntax in
separateTestModuleWithMinifiedApp example project.  The integration
base test involving that project previously "passed" because the only
test method asserted that there was a failure; however, the connected
test asserts that the connectedAndroidTest task is successful.

Bug: none filed
Test: existing
Change-Id: Ia59fc4ab36bfbcf5b0040abdedb16cb07c43ae1d