Run BuiltInKotlinForScreenshotTestTest with and without built-in Kotlin
Screenshot tests and test fixtures should work with both built-in Kotlin
and the legacy `kotlin-android` plugin.

Previously, BuiltInKotlinForScreenshotTestTest was running with only
builtInKotlin=false. This change makes the test run with
builtInKotlin=true too (similar to BuiltInKotlinForTestFixturesTest).

Also deletes BuiltInKotlinForAppTest.testWithScreenshotTestEnabled as
that test was testing screenshot tests when builtInKotlin=true, which is
now covered by the updated BuiltInKotlinForScreenshotTestTest.

Bug: 385745419
Test: Updated BuiltInKotlinForTestFixturesTest
Change-Id: Ie55a40e014f8721491bc720153a6108e83bd9bd6