[test] Test multi-module task cacheability
We don't appear to test the task states of library dependency projects
when they are built as part of an app build.

This change aims to track improvements and regressions for this case
eg. encoding relative resources to improve processResources cacheabilty
in I584fde15fa2b8a79e209cfb62ff440eaff4644f0.

Bug: 467645284
Test: LibraryCacheablityTest
Change-Id: I50d5cde3619e6ef5a934c430a75bdb8a9125d416